pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "questionarie"
version = "0.1.0"
description = "پرسشنامه ارزیابی ریسک سرمایه‌گذاری"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.31.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.1",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.1.1",
    "ruff>=0.2.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
artifacts = [
    "static/*",
    "data/*",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
